\name{muste}
\alias{muste}
\alias{muste-package}
\title{Launch Muste}
\description{
  Initializes Muste (an open-source initation of Survo) and opens the editorial interface window.
  See \href{http://www.survo.fi/muste/}{Muste web pages} and
  \href{http://www.survo.fi/english/}{Survo web pages} for more information.
  Especially \href{http://www.survo.fi/tour/}{getting started tour},
      \href{http://www.survo.fi/books/}{Survo books},
      \href{http://www.survo.fi/demos/}{Survo demos} and
      \href{http://www.survo.fi/help/}{Survo help system}
  are likely to be useful for a newbie. Please note, however, that the functionality of
  Muste is still quite limited if compared to Survo.
}
\usage{
  muste()
}